refactor(step-6): move notes.ts to src/operations/ and fold in 3 helpers
notes.ts joins bear-encoding.ts in the operations/ layer. As part of
the move, three helpers that lived in the grab-bag utils.ts get folded
in where they belong:
- parseDateString — already used only by searchNotes' date filter
- stripLeadingHeader and noteHasHeader — pure markdown content helpers
  used by handleNoteTextUpdate

Their 3 describe blocks move verbatim from utils.test.ts into a new
co-located operations/notes.test.ts, preserving all 17 assertions.
utils.test.ts is now empty of useful content and is deleted.

utils.ts shrinks to ~85 lines — it now contains only handleNoteTextUpdate
(which will move into tools/note-tools.ts in step 9, at which point
utils.ts disappears entirely). Its import of getNoteContent and the
two markdown helpers now resolves to ./operations/notes.js.

Tests: 37 total, same as before. Test files: utils.test.ts is gone
(19 tests → 18 moved to operations/notes.test.ts + 1 moved to
operations/bear-encoding.test.ts in step 3).
