+++ src/main.ts
@@ -13,7 +13,12 @@ import { handleNoteTextUpdate } from './utils.js';
 import { logger } from './logging.js';
 import { cleanBase64 } from './operations/bear-encoding.js';
 import { createErrorResponse, createToolResponse } from './tools/responses.js';
-import { awaitNoteCreation, findNotesByTitle, getNoteContent, searchNotes } from './notes.js';
+import {
+  awaitNoteCreation,
+  findNotesByTitle,
+  getNoteContent,
+  searchNotes,
+} from './operations/notes.js';
 import { findUntaggedNotes, listTags } from './tags.js';
 import { buildBearUrl, executeBearXCallbackApi } from './infra/bear-urls.js';
 import type { BearTag } from './types.js';
